Add logging to file and terminal
Check for verbosity and if the module is the current module
