+++ core/src/bmi.jl
@@ -614,38 +614,9 @@ BMI.get_end_time(model::Model) = seconds_since(model.config.endtime, model.confi
 BMI.get_time_units(model::Model) = "s"
 BMI.get_time_step(model::Model) = get_proposed_dt(model.integrator)
 
-"""
-    run(config_file::AbstractString)::Model
-    run(config::Config)::Model
-
-Run a [`Model`](@ref), given a path to a TOML configuration file, or a Config object.
-Running a model includes initialization, solving to the end with `[`solve!`](@ref)` and writing results with [`BMI.finalize`](@ref).
-"""
-run(config_file::AbstractString)::Model = run(Config(config_file))
-
-function is_current_module(log)
-    (log._module == @__MODULE__) ||
-        (parentmodule(log._module) == @__MODULE__) ||
-        log._module == OrdinaryDiffEq  # for the progress bar
-end
-
 function run(config::Config)::Model
-    global_logger(TeeLogger(FileLogger("ribasim.log"), TerminalLogger()))
-    logger = current_logger()
-
-    # Reconfigure the logger if necessary with the correct loglevel
-    # but make sure to only log from Ribasim
-    if min_enabled_level(logger) + 1 != config.logging.verbosity
-        logger = EarlyFilteredLogger(
-            is_current_module,
-            LevelOverrideLogger(config.logging.verbosity, logger),
-        )
-    end
-
-    with_logger(logger) do
-        model = Model(config)
-        solve!(model)
-        BMI.finalize(model)
-        return model
-    end
+    model = Model(config)
+    solve!(model)
+    BMI.finalize(model)
+    return model
 end
